requirements.txt
fastapi
uvicorn[standard]
telebot
pyTelegramBotAPI
openai
python-dotenv
httpx
fastapi[test]
